requirements.txt
python-telegram-bot==20.6
fastapi==0.104.1
uvicorn==0.23.2
sqlalchemy==2.0.23
python-dotenv==1.0.0
pytube==15.0.0
instaloader==4.10.1
yt-dlp==2023.10.13
shazamio==0.4.0.1
requests==2.31.0
beautifulsoup4==4.12.2
pyjwt==2.8.0
redis==5.0.1
aiohttp==3.10.11
asyncio==3.4.3